--- src/main/java/com/grewmeet/dating/datingcommandservice/dto/event/DatingMeetingCreated.java
@@ -1,13 +1,0 @@
-package com.grewmeet.dating.datingcommandservice.dto.event;
-
-import java.time.LocalDateTime;
-
-public record DatingMeetingCreated(
-        Long datingMeetingId,
-        String title,
-        String description,
-        LocalDateTime meetingDateTime,
-        String location,
-        Integer maxParticipants,
-        LocalDateTime createdAt
-) {}
